third_party/WebKit/LayoutTests/fast/doctypes/xml-doctype.xhtml

<!DOCTYPE HTML>
<html xmlns="http://www.w3.org/1999/xhtml">
<head>
    <script>
        if (window.testRunner)
            testRunner.dumpAsText();

        log = function(msg)
        {
            document.getElementById('console').appendChild(document.createTextNode(msg + "\n"));
        }

        onload = function()
        {
            if (document.doctype)
                log("PASS: a doctype in a XML document is not ignored.");
            else
                log("FAIL: a doctype in a XML document is ignored.");

            if (document.firstChild == document.doctype)
                log("PASS: a doctype in an XML document is correctly set to the first child of the Document.");
            else
                log("FAIL: a doctype in an XML document is not correctly set to the first child of the Document.");

            if (document.doctype.publicId == '')
                log("PASS: doctype.publicId correctly defaults to empty string when not set.");
            else
                log("FAIL: doctype.publicId does not default to empty string when not set, instead is : " + document.doctype.publicId);

            if (document.doctype.systemId == '')
                log("PASS: doctype.systemId correctly defaults to empty string when not set.");
            else
                log("FAIL: doctype.systemId does not default to empty string when not set, instead is : " + document.doctype.systemId);
        }
    </script>
</head>
<body>
    This tests the behavior of a Doctype in an XML document.
    <pre id="console"></pre>
</body>
</html>
